Check existing response before re-fetching
This checks if there is an existing OCSP response on disk for the cert
lineage, which doesn't expire for at least two days, before deciding to re-fetch a
response. This check is skipped when the script is run as a deploy hook,
because in that case the existing OCSP response would be for a previous
certificate in the lineage.

It also keeps track of how many responses are fetched in a run (if any),
and informs the user of this in the output.

Fixes #14.