Fix a few instances of a Bash pitfall
Two command substituted variables were set to read-only, but this masked
the return value, which was a bigger problem.

For the expiry check of the existing OCSP response, it's actually ok,
since it would just silently skip the check and still fetch the
response, which might be better than failing hard.