Install linters from Snap in CI
The `max_line_length` directive in EditorConfig doesn't seem to be supported
any more since shfmt 3.5.0. Can't find a proper source on this, but testing
seems to confirm it.

The above, and/or another change in shfmt, results in quite some lines needing
style changes.

Installing shfmt from Snap also means it can't be run as root any more, which
wasn't necessary or good anyway.